# APIserver/.docs/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: Mackyson
  description: HaiqueのもろもろのAPI
  license:
    name: MIT
  title: Haique
  version: "1.0"
servers:
- description: localhost
  url: http://localhost:8080
paths:
...
  /api/subscribe/{user_id}:
    delete:
      description: リムーブ
      operationId: delete-subscribe
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/inline_object_4'
        content:
          application/json:
            schema:
              properties:
                session_id:
                  format: uuid
                  type: string
              required:
              - session_id
              type: object
              x-examples:
                example-1:
                  session_id: hoge
      responses:
        "200":
          description: OK
    post:
      description: フォロー
      operationId: post-subscribe
      parameters:
      - explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/inline_object_3'
        content:
          application/json:
            schema:
              properties:
                session_id:
                  format: uuid
                  type: string
              required:
              - session_id
              type: object
              x-examples:
                example-1:
                  session_id: hoge
      responses:
        "200":
          description: OK
...
components:
  requestBodies:
...
    inline_object_4:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object_4'
    inline_object_3:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object_3'
...
  schemas:
...
    inline_object_3:
      properties:
        session_id:
          format: uuid
          type: string
      required:
      - session_id
      type: object
      x-examples:
        example-1:
          session_id: hoge
    inline_object_4:
      properties:
        session_id:
          format: uuid
          type: string
      required:
      - session_id
      type: object
      x-examples:
        example-1:
          session_id: hoge
...